LayoutTests/inspector/debugger/debugger-change-variable-expected.txt

Tests that modifying local variables works fine.

Debugger was enabled.
Set timer for test function.
Script execution paused.
Evaluated script on the calling frame: 310
Set property value error: Failed to change variable value: TypeError: Variable value mutation is not supported
Evaluated script on the calling frame (after value modification): 310
Script execution resumed.
Debugger was disabled.
